Add in battery allocation prototype and fixes for observer (ROSI-AP/rosi-ap_commercial/cag/rmf_scheduler!32)
* Lint

* Fix unit tests


* Update scheduler


* Added utility method for datetime string and modified telegram messages

* Update charging point

* Change charging waypoint


* Re-enable optimization and change charge task to go to place


* temp

* Quick changes

* Change charger temporarily

* some fixes

* Small fixes on automatic schedule deletion


* Add notification for frequency not met


* Enable duration live updates


* Add in swap battery dirty fixes


* Avoid JSON parsing error from task states


* Better runtime update for DAG and dangling events


* Remove some print statements


* Enable http client keycloak authentication


* Fix build failure for execution constructor


* Force completion callback to tick


* Automate optimization and deconfliction


* lint and fix test


* Implement caching in scheduler and add test case


* Add local caching capability


* Fix completion delay crash


* Skip unknown type for scheduler api


* Add completion detail handling


* temp

* Reuse old plugin loader pointer


* Fix executor failure due to empty task details

+++ rmf_scheduler_plugins/src/robot_task_estimate_client.cpp
@@ -89,7 +89,7 @@ RobotTaskEstimateClient::async_estimate(
     state["waypoint"] = request.state->waypoint;
     state["orientation"] = request.state->orientation;
     state["battery_soc"] = request.state->consumables["battery_soc"];
-    state["time"] = request.start_time / 1e6;
+    state["time"] = static_cast<uint64_t>(request.start_time / 1e6);
     task_request["state"] = state;
   }
